change dataType id in product_add_edit

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/fragment/AddEditProductPreviewFragment.kt
@@ -160,7 +160,7 @@ class AddEditProductPreviewFragment :
     private var latitude: String = ""
     private var longitude: String = ""
     private var postalCode: String = ""
-    private var districtId: Int = 0
+    private var districtId: Long = 0
     private var formattedAddress: String = ""
     private var productInputModel: ProductInputModel? = null
     private var isFragmentVisible = false
@@ -1557,7 +1557,7 @@ class AddEditProductPreviewFragment :
     private fun getSaveShopShippingLocationData(
             shopId: Int,
             postCode: String,
-            courierOrigin: Int,
+            courierOrigin: Long,
             addrStreet: String,
             lat: String,
             long: String
@@ -1593,7 +1593,7 @@ class AddEditProductPreviewFragment :
                 postalCode.isNotBlank() &&
                 latitude.isNotBlank() &&
                 longitude.isNotBlank() &&
-                districtId != 0 &&
+                districtId != 0L &&
                 formattedAddress.isNotBlank()) {
 
             val params = getSaveShopShippingLocationData(
